requirements.txt
flask>=3.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
requests>=2.31.0
werkzeug>=3.0.0
